requirements.txt
openai>=1.12.0
requests>=2.31.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
